+++ Assets/Scripts/CameraController.cs
@@ -1,8 +1,7 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.EventSystems;
 
-public class CameraController : MonoBehaviour
+public class CameraController : MonoBehaviour, IPointerClickHandler, IPointerDownHandler, IPointerUpHandler
 {
     public GameObject player;
     private Vector3 camOffset;
@@ -11,11 +10,36 @@ public class CameraController : MonoBehaviour
     void Start()
     {
         camOffset = transform.position - player.transform.position;
+        addPhysicsRaycaster();
+    }
+
+    public void addPhysicsRaycaster()
+    {
+        PhysicsRaycaster raycaster = GameObject.FindObjectOfType<PhysicsRaycaster>();
+        if (raycaster != null)
+        {
+            Camera.main.gameObject.AddComponent<PhysicsRaycaster>();
+        }
     }
 
     // Update is called once per frame
     void LateUpdate()
     {
         transform.position = player.transform.position + camOffset;
+    }
+
+    public void OnPointerDown(PointerEventData eventData)
+    {
+        Debug.Log("Clicked: " + eventData.pointerCurrentRaycast.gameObject.name);
+    }
+
+    public void OnPointerUp(PointerEventData eventData)
+    {
+        Debug.Log("Clicked: " + eventData.pointerCurrentRaycast.gameObject.name);
+    }
+
+    public void OnPointerClick(PointerEventData eventData)
+    {
+        Debug.Log("Clicked: " + eventData.pointerCurrentRaycast.gameObject.name);
     }
 }